requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
ratelimit==2.2.1
cached-property==1.5.2
# If you want Selenium fallback for JS-heavy pages:
selenium==4.12.0
webdriver-manager==4.0.2
